Map LCVR voltage changes to corresponding epoch
In the event of a queued stream of epochs in the GENLOG pipe for polarization compensation, the next QBER calculated during the next epoch does not correspond to the changed active polarization compensation values in LCVR. Current (naive) strategy is to discard all such currently queued epochs.

In practice, this queueing does not occur during local QKD deployment tests with 600 coincidences/epoch and QBER < 11%.

A different strategy ought to be used when employing the stochastic minimization algorithm that runs over several LCVR values concurrently.