Change base image to python3.9-slim
Rebuilding of images tend to incur new dependency-compatibility problems, especially that of matplotlib, when the base image is python3.9-alpine. This is because the libraries have to be built from source, and seems to be triggered because there are no native binaries built with the underlying 'musl' C library[1].

Changing to the Debian base improved image build time by an order of magnitude (~30 mins to ~1m), as well as reduced the image size (~1 GB to 880 MB).

[1]: https://pythonspeed.com/articles/alpine-docker-python/