Enable feedback to freqcd only after backlog
Previous implementation caused an overcompensation of clock skew
due to the backlog of epochs to be processed by costream, which will not
be retroactively adjusted for frequency differences (which would involve
another intermediary freqcd).

Instead, we hope that the frequency drift is not too strong in a single
direction - while fpfind is working its magic - and allow epoch feedback
only after the current epoch at time of recording has passed.