Use sliding mean for QBER minimization
Previous implementation uses an absolute minimum for convergence, which
resulted in observations of QBER not being actively compensated after
some time, due to the low QBER arising from statistical fluctuations.
This is less noticeable when the overall QBER is relatively stable.

QBER history is used to allow the target QBER to shift back towards the
mean of observed QBER, so that the polarization compensation can
effectively try to push it down again. This is encapsulated in the
'qber_history_length' configuration value.