Allow server to exit gracefully upon termination
QKDServer previously does not terminate, since the SIGTERM signal sent
by 'docker stop' and 'docker rm' is never passed to Gunicorn (instead
caught and ignored by the entrypoint bash script). This necessitated a
timeout to wait for Docker to forcibly teardown the container, even
after a 'docker rm'.

This set of changes implements:

  - SIGTERM passing to Gunicorn
  - Gunicorn triggering QKDController to stop/kill its processes
  - QKDController exiting with ecode 0/1 respectively
  - Gunicorn has a 5s timeout for SIGTERM, before sending SIGKILL