fix: ensure polcomp optimizer init before call
Previously a possible race condition existed when a successive epoch is
passed to polarization compensation, which may coincide with an ongoing
optimizer refresh, since both processes are running asynchronously. This
may in turn cause the thread to freeze.

Added some debugging logs when passed epochs are rejected, as well as
aborting if such a race condition has been triggered.